refactor: move initialize_agents to src/engine/setup.py (closes #57)
