+++ doc/DetailDesign.md
@@ -11,6 +11,7 @@
 | モジュール | 責務 |
 |---|---|
 | `game.py` | ゲーム全体の状態マシン。昼夜ループを回す |
+| `setup.py` | ゲーム初期化（エージェント生成・役職割り当て・永続化） |
 | `phase.py` | フェーズ定義（昼/夜/開始/終了）と遷移ロジック |
 | `vote.py` | 投票集計・同数処理・追放決定 |
 | `victory.py` | 勝利条件の判定 |
