+++ main.py
@@ -11,11 +11,8 @@ Usage:
     uv run main.py --replay --spectator      # Replay mode (spectator)
 """
 import argparse
-import json
 import os
-import random
 import sys
-from pathlib import Path
 
 from dotenv import load_dotenv
 
@@ -26,55 +23,10 @@ if not os.environ.get("ANTHROPIC_API_KEY"):
     print("Copy .env.example to .env and add your API key.")
     sys.exit(1)
 
-from src.domain.agent import AgentState, Persona, Belief  # noqa: E402
-from src.agent import store  # noqa: E402
+from src.engine.setup import initialize_agents  # noqa: E402
 from src.engine.game import GameEngine  # noqa: E402
 from src.logger.writer import LogWriter, archive_state  # noqa: E402
 from src.ui.cli import CLI  # noqa: E402
-
-
-def initialize_agents(num_players: int) -> list[AgentState]:
-    """Create and persist initial agent states with randomized roles."""
-    Path("state/agents").mkdir(parents=True, exist_ok=True)
-
-    agent_configs = json.loads(Path("config/agents.json").read_text(encoding="utf-8"))
-    roles_config = json.loads(Path("config/roles.json").read_text(encoding="utf-8"))
-
-    key = str(num_players)
-    if key not in roles_config:
-        print(f"Error: no role configuration found for {num_players} players.")
-        print(f"Available: {', '.join(roles_config.keys())} players")
-        sys.exit(1)
-
-    roles = roles_config[key]
-    selected_configs = agent_configs[:num_players]
-    if len(selected_configs) < num_players:
-        print(f"Error: not enough agents in config/agents.json for {num_players} players (found {len(agent_configs)}).")
-        sys.exit(1)
-
-    shuffled_roles = roles[:]
-    random.shuffle(shuffled_roles)
-
-    agents = []
-    for config, role in zip(selected_configs, shuffled_roles):
-        name = config["name"]
-        beliefs = {
-            other["name"]: Belief()
-            for other in selected_configs
-            if other["name"] != name
-        }
-        agent = AgentState(
-            name=name,
-            role=role,
-            persona=Persona.model_validate(config),
-            beliefs=beliefs,
-            memory_summary=[],
-            is_alive=True,
-        )
-        store.save(agent)
-        agents.append(agent)
-
-    return agents
 
 
 def main() -> None:
